Feito substituicao das classes de SalesAdBuyer por ItemOfProduct e refatorados os seus metodos e atributos.

+++ src/main/java/com/mercadolibre/demo/model/ItemOfProduct.java
@@ -18,21 +18,19 @@ import javax.persistence.*;
 public class ItemOfProduct implements Serializable {
 
 	private static final long serialVersionUID = 1L;
-
-
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // Gera autoincrimento no banco de dados
     @Column(name = "idItem_of_product")
     private Long id;
 
     @Column(name = "quantity")
-    private int quantity;
+    private Long quantity;
 
     @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     @JoinColumn(name = "idsales_ad")
     private SalesAd salesAd;
 
-    public ItemOfProduct(int quantity, SalesAd salesAd) {
+    public ItemOfProduct(Long quantity, SalesAd salesAd) {
         this.quantity = quantity;
         this.salesAd = salesAd;
     }
